requirements_minimal.txt
# Core Framework
Flask==3.0.0
WTForms==3.1.1
Flask-WTF==1.2.1

# Web Scraping & Parsing
feedparser==6.0.11
beautifulsoup4==4.12.2
requests==2.31.0

# AI Integration
openai==1.6.1

# Task Scheduling
APScheduler==3.10.4

# Email
sendgrid==6.11.0

# Environment & Configuration
python-dotenv==1.0.0

# Additional Dependencies
Werkzeug==3.0.1
MarkupSafe==2.1.3
Jinja2==3.1.2
click==8.1.7
soupsieve==2.5
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
urllib3==2.1.0